requirements.txt
flask>=3.0.0
google-genai>=1.0.0
python-dotenv
llama-index
pinecone-client
sentence-transformers
torch